Add APS1 support for LOAD/REPEAT instructions.
Adds a loop-unrolling pass to APS1 export. This pass attempts to determine if a
sequence can be repeated with the miniLL repeat parameter. If not, it unrolls
repeated sequences.

To make testing of this easier, I have modified the equality concept for
PulseBlocks to ignore labels, and fixed equality testing for Pulse(s).