fix for #69,#71, and #72
#69
In the Looper Form of MeasFilterViews.enaml, test that the selected
filter is in the filter dictionary else return -1

#71
In the QtListStrWidget, test if the enable changed when the labels are the same
and the label is in the item list. Also, only signal a name change when the
labels are not the same and the newlabel is not in the item list. In the old
version of code there are some conditions when on_edit is called that result in
the dictionary manager trying to rename an item that does not exist.

#72
Added a new feature to save config in a different directory and load a config
from a different directory.  After loading a config, the GUI will restart itself
for the changes to take effectively