Choose translator from file extension.
Decouples ChannelLibrary setup from sequence file names, which is especially
handy when debuging test vectors. To get this to work, we need to handle the
fact that the APS1 and APS2 both use an h5 file extension. So, this also adds an
`is_compatible_file(filename)` method which drivers must define.