Add a scheduler to parallelize serial pulse sequences.
Basic functionality works with a simple stream of single qubit gates and
measurements. Does not yet handle control flow. Initial work on 2-qubit gates
reveals a potential representation problem. This issue is that some gates like
CNOT_CR actually return a list. If we simply flatten this list and then proceed
like normal, than a sequence like:
    seq = [X(q1), CNOT(q2, q3), Y(q1)]

would cause the scheduler to place the Y(q1) "inside" the unrolled CNOT(q2, q3).
This suggests to me that we may need to introduce another representation for
gate sequences like CNOT_CR so that we can properly recognize them as a unit.