Add a CompoundGate type to the pulse hierarchy.
This is a thin wrapper around a python list that let's us define '*' to
broadcast across the elements of two lists. This will be used by things like
CNOT_CR so that they can live naturally inside a sequence, and so that we can
describe parallel CNOTs.

This also modifies the promotion rules for '*' on the various pulse types, so
that Pulse, CompositePulse, and PulseBlock will all promote to CompoundGate when
multiplied with another CompoundGate.

Not yet passing tests.