Go back to broadcasting all control flow.
To support more narrow control flow channel targets, we need a mechanism to
expand channel targets of Wait() and friends with other channels associated with
`Qubit`s. For instance,
	seq = [WAIT((q1,)), X(q1), MEAS(q1)]
really needs to be interpetted as
	seq = [WAIT((q1, M-q1)), X(q1), MEAS(q1)]
or even
	seq = [WAIT((q1, M-q1, q1-gate, M-q1-gate, digitizerTrig, slaveTrig)),
	       X(q1),
		   MEAS(q1)]
since those other channels will be in the sequence after other compiler passes
run.

Multi-qubit sequences expand the list even further such that we need to include
`Edge`s.

Until we have a mechanism that can do this, I am reverting to simple
broadcasting of all control flow.