move config_loc stuff; add tests
There's a problem with putting this inside the QGL module;
the __init__ for that module imports many things, and by
the time it is done it has imported all of the things whose
importing we're trying to override with this!  Therefore we
need to put it in its own area, where packages are not
blindly imported.