delay loading drivers until they're needed
if we try to do an import_module of the drivers
before everything else is loaded, we run into some
dependency issues.  We're abusing the module loader
and we pay the price if we try to load the driver
modules before their prereqs are available.  (I don't
understand exactly why, but I'm not going to fight it)